Initialize various pointers to NULL
A number of pointers in the HeatPump class were used before they were
initialized, particularly the callback function pointers. This becomes
an issue when an object is created from the class using heap memory, as
there is no guarantee that the memory will be initialized to all zeros
in that case.