Allow explicitly setting baud rate
The MSZ-GB50VA requires a baud rate of 9600. The existing code to
automatically retry at 9600 baud if 2400 fails does not work on this
unit -- the connection attempt at 9600 baud fails if there has been a
previous attempt at 2400 baud. I assume it's something to do with the
heat pump's internal serial buffering. Anyway, this commit allows you to
select the correct baud rate first time around if you know that it's not
the default, which solves the issue.

Previously you could call `connect(&Serial)` and it would try 2400
followed by 9600. This remains the same.

Previously you could call `connect(&Serial, false)` and it would only
try 2400. Now the equivalent call would be `connect(&Serial, 2400)`.

Previously you couldn't set the initial baud rate. Now you can call
`connect(&Serial, 9600)`.

Which is to say, there is a minor backwards-incompatible change to the
constructor API but it's unlikely to cause a problem for anyone.